Broke system logs into multiple components
