+++ frontend/src/pages/SystemLogs.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SystemLogs from '../components/Admin/System-Logs/SystemLogs';
+
+const SystemLogsPage: React.FC = () =>{
+    return (
+        <SystemLogs />
+    );
+};
+
+export default SystemLogsPage;
