+++ frontend/src/router/router.tsx
@@ -16,6 +16,7 @@ import Account from '../components/Account';
 import SingleBooking from '../components/Booking';
 import ManageUsers from '../components/Admin/ManageUsers';
 import SystemLogs from '../components/Admin/System-Logs/SystemLogs';
+import SystemLogsPage from '../pages/SystemLogs';
 
 export const router = createBrowserRouter([
   {
@@ -81,7 +82,7 @@ export const router = createBrowserRouter([
       },
       {
         path: "/admin/logs",
-        element: <SystemLogs />
+        element: <SystemLogsPage />
       },
       {
         path: "/account",
